# Conflicts:
#	lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene90/Lucene90HnswVectorsReader.java
#	lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene91/Lucene91HnswVectorsReader.java
#	lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene94/OffHeapFloatVectorValues.java
#	lucene/backward-codecs/src/test/org/apache/lucene/backward_codecs/lucene94/Lucene94HnswVectorsWriter.java
#	lucene/codecs/src/java/org/apache/lucene/codecs/simpletext/SimpleTextKnnVectorsReader.java
#	lucene/core/src/java/org/apache/lucene/codecs/KnnVectorsWriter.java
#	lucene/core/src/java/org/apache/lucene/codecs/perfield/PerFieldKnnVectorsFormat.java
#	lucene/core/src/java/org/apache/lucene/document/KnnFloatVectorField.java
#	lucene/core/src/java/org/apache/lucene/index/BufferingKnnVectorsWriter.java
#	lucene/core/src/java/org/apache/lucene/index/CheckIndex.java
#	lucene/core/src/java/org/apache/lucene/index/CodecReader.java
#	lucene/core/src/java/org/apache/lucene/index/VectorValues.java
#	lucene/core/src/java/org/apache/lucene/search/AbstractKnnVectorQuery.java
#	lucene/core/src/java/org/apache/lucene/util/hnsw/HnswGraphSearcher.java
#	lucene/core/src/java/org/apache/lucene/util/hnsw/RandomAccessVectorValues.java
#	lucene/core/src/test/org/apache/lucene/index/TestKnnGraph.java
#	lucene/core/src/test/org/apache/lucene/search/BaseKnnVectorQueryTestCase.java
#	lucene/core/src/test/org/apache/lucene/util/hnsw/KnnGraphTester.java
#	lucene/test-framework/src/java/org/apache/lucene/tests/codecs/asserting/AssertingKnnVectorsFormat.java

+++ lucene/core/src/test/org/apache/lucene/util/packed/TestPackedInts.java
@@ -703,7 +703,7 @@ public class TestPackedInts extends LuceneTestCase {
     }
 
     // test ramBytesUsed
-    assertEquals(RamUsageTester.ramUsed(writer), writer.ramBytesUsed(), 8.d);
+    assertEquals((double) RamUsageTester.ramUsed(writer), (double) writer.ramBytesUsed(), 8.d);
 
     // test copy
     PagedGrowableWriter copy =
